fix: move recruit detail sections out of app layer per fsd
the recruit/[idx] route segment carried widget-level ui (6 section
components), domain constants, utils, and a test. relocate per fsd:

- sections -> src/widgets/recruit/detail/<section>/ with a shared
  widget stylesheet (sections share section/list/notice_text classes)
- BENEFITS/HIRING_PROCESS -> entities/recruit/constants
- toIdx/formatDate -> entities/recruit/util/format (test follows)
- markdown component map -> shared/ui/markdown with generic class names

route now keeps only page shell files; client.tsx stays as the
section orchestrator with widget imports.
