+++ src/entities/recruit/constants/recruit-detail.constants.ts
