+++ src/widgets/recruit/detail/notice/index.tsx
